fix: secure query by design

+++ src/lambda-car-backend/commands/refueling_commands.py
@@ -2,11 +2,13 @@ from dataclasses import dataclass
 from uuid import UUID
 from datetime import datetime
 
+from decimal import Decimal
+
 @dataclass
 class CreateRefuelingCommand:
     car_id: UUID
     card_number: str
-    liter_price: float
+    liter_price: Decimal
     liters: int
     receipt_filename: str
     receipt_content: bytes
@@ -18,6 +20,6 @@ class UpdateRefuelingCommand:
     refueling_id: UUID
     car_id: UUID
     card_number: str
-    liter_price: float
+    liter_price: Decimal
     liters: int
     date: datetime
